guild: PT-B2 审计 R1-R7 修复——打包/字段布局/包映射/GetBlackListSize int/packetSize/cwtl/三元改分支，169->168

+++ dnf_decompile/source/DNFServer/GameServer/Guild/DNFUser.cpp
@@ -462,9 +462,9 @@ int CUser::IsBlackUser(unsigned int charNo)
     return it != m_blackList.end() ? 1 : 0;
 }
 
-unsigned short CUser::GetBlackListSize()
+int CUser::GetBlackListSize()
 {
-    return (unsigned short)m_blackList.size();
+    return (int)m_blackList.size();
 }
 
 void CUser::GuildInviteProcess()
